pyproject.toml
[tool.poetry]
name = "smart_evidence"
version = "0.1.0"
description = ""
authors = ["Oğuz Şerbetçi <oguz.serbetci@impactnexus.io>"]

[tool.poetry.dependencies]
python = "~3.10"
pyairtable = "^1.0"
spacy = { version = "^3.4.0" }
scikit-learn = "^1.0"
python-dotenv = "^0.20"
pycld2 = "^0.41"
opensearch-py = "^2.0.0"
requests-aws4auth = "^1.1"
boto3 = "^1.18"
sentence-transformers = "^2.2.2"
farm-haystack = "^1.7.0"
huggingface-hub = "<0.8.0"
transformers = "4.20.1"
setfit = "0.3.0"
pandas = "<1.4.0"
numba = "^0.55"
numpy = "<1.23"

fastapi = "^0.75"
uvicorn = "^0.17"
gunicorn = "^20.1"
fastapi-auth0 = "^0.3.0"

chainer = { version = '*', optional = true }
rdflib = { version = "^6.1.1", optional = true }
titlecase = { version = "*", optional = true }
keybert = { version = "^0.5.0", optional = true }
pke = { git = "https://github.com/boudinfl/pke.git", optional = true }
openpyxl = { version = "^3.0.9", optional = true }
nltk = { version = "^3.6.7", optional = true }
seaborn = { version = "^0.11.2", optional = true }
pytextrank = { version = "^3.2.2", optional = true }
plotly = { version = "^5.5.0", optional = true }
sagemaker = { version = "*", optional = true }
scattertext = { version = "^0.1.5", optional = true }
neomodel = "^4.0.8"
spacy-dbpedia-spotlight = "^0.2.2"
py2neo = "^2021.2.3"
thefuzz = "^0.19.0"
trafilatura = "^1.4.0"
"pdfminer.six" = "^20220524"
python-levenshtein = "^0.20.8"

[tool.poetry.group.dev.dependencies]
black = "^22.6"
ipykernel = "^6.6.0"
nbformat = "^5.1.3"
mypy = "^0.991"
twine = "^3.8.0"
pytest = "^7.1.3"
types-boto3 = "^1.0.2"

[tool.poetry.extras]
notebooks = [
    'titlecase',
    'rdflib',
    'pke',
    'nltk',
    'openpyxl',
    'nltk',
    'seaborn',
    'pytextrank',
    'plotly',
]
pjobs = ['sagemaker']
gpu = ["chainer", "cupy-cuda114"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
